+++ jobslib/oneinstance/consul.py
@@ -20,20 +20,27 @@ __all__ = ['ConsulLock']
 
 logger = logging.getLogger(__name__)
 
+ONE_DAY_SECONDS = 60 * 60 * 24
+
 
 class ConsulLock(BaseLock):
     """
-    Consul lock implementation. Provides locking among datacenters.
-    When the lock expires due to TTL, :exc:`OneInstanceWatchdogError` is
-    raised.
+    Consul lock implementation, provides locking among datacenters.
+    When the lock expires due to TTL, :exc:`OneInstanceWatchdogError`
+    is raised. It is possible to extend the lock using :meth:`refresh`.
+    Lock is not extended immediately, but request for extending is
+    made and lock will be extended asynchronously when **SIGALRM** is
+    received. So use conservative TTL and periodically extend the lock.
 
     .. warning::
 
         :class:`ConsulLock` uses :mod:`signal` and :data:`signal.SIGALRM`
         for TTL mechanism, so don't use :data:`signal.SIGALRM` in your
-        task.
+        task. And don't use multiple instances of the :class:`ConsulLock`
+        at the same time, becase **SIGALRM** can't be shared among
+        multiple instances.
 
-    For using the :class:`ConsulLock` write into **settings**:
+    For using the :class:`ConsulLock` configure backend in **settings**:
 
     .. code-block:: python
 
@@ -42,6 +49,7 @@ class ConsulLock(BaseLock):
             'options': {
                 'key': 'jobs/example/lock',
                 'ttl': 60,
+                'lock_delay': 15,
             },
         }
     """
@@ -61,51 +69,42 @@ class ConsulLock(BaseLock):
                 return key
             return self._settings['key']
 
-        @option(attrtype=int)
+        @option(required=True, attrtype=int)
         def ttl(self):
             """
-            Maximum lock lifespan in seconds.
+            Maximum lock lifespan in seconds, must be between 10 seconds
+            and one day. If value is omitted, default is one day (maximum
+            for Consul).
             """
-            one_day_seconds = 60 * 60 * 24
-
-            ttl = os.environ.get('JOBSLIB_ONE_INSTANCE_OPTIONS_TTL')
-            if ttl:
-                ttl = int(ttl)
+            if 'JOBSLIB_ONE_INSTANCE_OPTIONS_TTL' in os.environ:
+                ttl = int(os.environ['JOBSLIB_ONE_INSTANCE_OPTIONS_TTL'])
             else:
-                ttl = self._settings.get('ttl', one_day_seconds)
-            if ttl < 10 or ttl > one_day_seconds:
+                ttl = self._settings.get('ttl', ONE_DAY_SECONDS)
+            if ttl < 10 or ttl > ONE_DAY_SECONDS:
                 raise ValueError(
                     'TTL must be between 10 and {} seconds'.format(
-                        one_day_seconds))
+                        ONE_DAY_SECONDS))
             return ttl
 
-        @option(attrtype=int)
+        @option(required=True, attrtype=int)
         def lock_delay(self):
             """
-            When sessions invalidation request is received, wait *lock_delay*
-            seconds before session is truly invalidated. Valid value must be
-            between 0 and 60 seconds, default is 1.
+            When sessions invalidation request is received, wait
+            *lock_delay* seconds before session is truly invalidated.
+            Value must be between 0 and 60 seconds, default is 1.
             """
-            lock_delay = os.environ.get('JOBSLIB_ONE_INSTANCE_LOCK_DELAY')
-            if lock_delay:
-                lock_delay = int(lock_delay)
+            if 'JOBSLIB_ONE_INSTANCE_LOCK_DELAY' in os.environ:
+                delay = int(os.environ.get('JOBSLIB_ONE_INSTANCE_LOCK_DELAY'))
             else:
-                lock_delay = self._settings.get('lock_delay', 1)
-            if lock_delay < 0 or lock_delay > 60:
+                delay = self._settings.get('lock_delay', 1)
+            if delay < 0 or delay > 60:
                 raise ValueError('lock_delay must be between 0 and 60 seconds')
-            return lock_delay
+            return delay
 
     def __init__(self, context, options):
         super().__init__(context, options)
-        self.session_id = None
-
-    @staticmethod
-    def alarm_handler(unused_signum, unused_frame):
-        """
-        Raise :exc:`OneInstanceWatchdogError` when TTL of the lock is
-        reached.
-        """
-        raise OneInstanceWatchdogError
+        self._session_id = None
+        self._refresh_lock_flag = False
 
     def acquire(self):
         timestamp = get_current_time()
@@ -123,48 +122,60 @@ class ConsulLock(BaseLock):
                 self.options.key, ujson.dumps(record), acquire=session_id)
         except Exception:
             logger.exception("Can't acquire lock")
-            self.context.consul.session.destroy(session_id)
-            return False
         else:
             if res is True:
-                self.session_id = session_id
-                # Set SIGALRM handler. If lock is not released before ttl,
-                # process will be killed.
-                signal.signal(signal.SIGALRM, self.alarm_handler)
+                self._session_id = session_id
+                self._refresh_lock_flag = False
+                # Set SIGALRM refresh handler. If lock is not released or
+                # extended before ttl is reached, task will be killed.
+                signal.signal(signal.SIGALRM, self._alarm_handler)
                 signal.alarm(self.options.ttl)
                 return True
-
-            self.context.consul.session.destroy(session_id)
-            return False
+            logger.error("Can't acquire lock")
+        self.context.consul.session.destroy(session_id)
+        return False
 
     def release(self):
         try:
             res = self.context.consul.kv.put(
-                self.options.key, None, release=self.session_id)
+                self.options.key, None, release=self._session_id)
         except Exception:
             logger.exception("Can't release lock")
-            return False
         else:
-            self.session_id = None
             if res is True:
+                self._session_id = None
+                self._refresh_lock_flag = False
                 # Cancel SIGALRM
                 signal.alarm(0)
                 signal.signal(signal.SIGALRM, signal.SIG_DFL)
                 return True
-            return False
+            logger.error("Can't release lock")
+        return False
 
-    def refresh(self, ttl=None):
-        try:
-            res = self.context.consul.session.renew(self.session_id)
-        except Exception:
-            logger.exception("Can't refresh lock")
-            return False
-        else:
-            if res:
-                # Restart SIGALRM
-                signal.alarm(ttl if ttl is not None else self.options.ttl)
-                return True
-            return True
+    def refresh(self):
+        self._refresh_lock_flag = True
+        return True
+
+    def _alarm_handler(self, unused_signum, unused_frame):
+        """
+        **SIGALRM** signal handler, it is called at the end
+        of Consuls' session TTL. Extend the lock if request
+        for extending is presented, otherwise raise
+        :exc:`OneInstanceWatchdogError`.
+        """
+        if self._refresh_lock_flag:
+            try:
+                res = self.context.consul.session.renew(self._session_id)
+            except Exception:
+                logger.exception("Can't extend lock")
+            else:
+                if res:
+                    self._refresh_lock_flag = False
+                    # Restart SIGALRM
+                    signal.alarm(self.options.ttl)
+                    return
+                logger.error("Can't extend lock")
+        raise OneInstanceWatchdogError
 
     def get_lock_owner_info(self):
         unused_index, res = self.context.consul.kv.get(self.options.key)